Add KEntity class and moved entity.cpp files into it.

+++ include/enemyc.h
@@ -1,6 +1,6 @@
 #pragma once
 
-int select_encounter(int, int);
+size_t select_encounter(int, int);
 void enemy_init(void);
 void enemy_chooseaction(size_t);
 void enemy_charmaction(size_t);
